[Delegate] Converted delegate to expect f32 inputs (#325)
By default, the delegate now expects f32 inputs instead of bf16.

For the Large Matmul case, there are now two PDLs for rewriting to call
the delegate, one that matches bf16 inputs (large-matmul.pdl.mlir) and
one that matches f32 inputs (large-matmul-f32.pdl.mlir).

+++ experimental/delegate/mlp_aie_bf16_plugin.cpp
@@ -109,8 +109,8 @@ using B_DATATYPE = bfloat16_t;
 using C_DATATYPE = float; // bfloat16_t;
 
 // Types of the matmul LHS, RHS, and result, as seen by the model
-using ModelLhsDType = bfloat16_t;
-using ModelRhsDType = bfloat16_t;
+using ModelLhsDType = float; // bfloat16_t;
+using ModelRhsDType = float; // bfloat16_t;
 using ModelReturnDType = float;
 
 // Set to 1 if the kernel requires a pre-initialized buffer to be loaded
